Initial setup
We've got the basic mock git structure here of a repo, branches and commits. In
this context a commit is no more than a sha in a branch.

We've also introduced random timing for branching off and creating commits. It's
all in a single thread at this point as we only have one branch, but we'll need
threading later when we have more branches.