Simplify service references
We were passing too many references around, and we are about to introduce a new
service to be able to keep track of the stats. Instead we have singleton
instances of the services, and a single config class to simplify the
constructors.

For some reason we need to make sure that we initialise the GitClient before
the Printer. I'm not sure why, as the printer should lazily instantiate the git
service when needed, but there's probably a circular reference, or maybe a
threading issue. Whatever it is this works as-is.