Use simulated time
As we increased the time speed we were getting bigger and bigger discrepencies
due to the time it takes to do real Git operations. These should be treated as
instant when compared to the time to merge branches and run CI. By creating our
own timer we can pause time when doing any Git.